+++ src/ahbicht/expressions/expression_resolver.py
@@ -3,11 +3,16 @@ This module makes it possible to parse expressions including all their subexpres
 for example ahb_expressions which contain condition_expressions or condition_expressions which contain packages.
 Parsing expressions that are nested into other expressions is refered to as "resolving".
 """
+from typing import Type
 
-from lark import Transformer, Tree
+import inject
+from lark import Token, Transformer, Tree
+from lark.exceptions import VisitError
 
 from ahbicht.expressions.ahb_expression_parser import parse_ahb_expression_to_single_requirement_indicator_expressions
+from ahbicht.expressions.base_transformer import BaseTransformer
 from ahbicht.expressions.condition_expression_parser import parse_condition_expression_to_tree
+from ahbicht.expressions.package_expansion import PackageResolver
 
 
 def parse_expression_including_unresolved_subexpressions(expression: str, resolve_packages: bool = False) -> Tree:
@@ -33,6 +38,18 @@ def parse_expression_including_unresolved_subexpressions(expression: str, resolv
     return resolved_expression_tree
 
 
+async def expand_packages(parsed_tree: Tree) -> Tree:
+    """
+    Replaces all the "short" packages in parser_tree with the respective "long" condition expressions
+    """
+    try:
+        result = PackageExpansionTransformer().transform(parsed_tree)
+    except VisitError as visit_err:
+        raise visit_err.orig_exc
+
+    return result
+
+
 # pylint: disable=no-self-use, invalid-name
 class AhbExpressionResolverTransformer(Transformer):
     """
@@ -45,3 +62,18 @@ class AhbExpressionResolverTransformer(Transformer):
         """
         condition_tree = parse_condition_expression_to_tree(expression.value)
         return condition_tree
+
+
+# pylint: disable=no-self-use, invalid-name
+class PackageExpansionTransformer(Transformer):
+    """
+    The PackageExpansionTransformer expands packages inside a tree to condition expressions by using a PackageResolver.
+    """
+
+    def __init__(self):
+        super().__init__()
+        self._resolver: PackageResolver = inject.instance(PackageResolver)
+
+    async def package(self, token: Token):
+        ce = await self._resolver.get_condition_expression(token.value)
+        return parse_condition_expression_to_tree(ce)
