+++ unittests/test_expression_resolver.py
@@ -96,6 +96,7 @@ the following characters: [ ] ( ) U ∧ O ∨ X ⊻ and digits.""" in str(
 
         assert """Please make sure that:
              * all conditions have the form [INT]
+             * all packages have the form [INTP]
              * no conditions are empty
              * all compositions are combined by operators 'U'/'O'/'X' or without an operator
              * all open brackets are closed again and vice versa
